# server/standards/openapi.yaml
openapi: 3.0.0
info:
  title: IFC Processing Server API
  description: REST API for IFC file processing and component querying with in-memory indexing
  version: 0.1.0
  contact:
    name: IFC Processing Team
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.example.com
    description: Production server
...
paths:
...
  /api/upload:
    post:
      tags:
        - File Management
      summary: Upload and process IFC or JSON file
      description: |
        Upload an IFC (Industry Foundation Classes) or JSON file for processing.
        IFC files are automatically converted to JSON format and stored.
        Maximum file size is 500MB.
      operationId: uploadFile
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: The IFC or JSON file to upload
              required:
                - file
      responses:
        '200':
          description: File successfully processed and stored
          content:
            application/json:
              schema:
                type: object
                properties:
                  filename:
                    type: string
                    example: HelloWallIFCjsonC-2x3.json
                  entities_count:
                    type: integer
                    example: 42
                  stored_count:
                    type: integer
                    example: 42
                  store_path:
                    type: string
                    example: /path/to/datastores/HelloWallIFCjsonC-2x3
                  message:
                    type: string
                    example: Successfully processed 42 entities
        '400':
          description: Invalid input (missing file, invalid format, or invalid JSON)
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: 'No file provided'
        '413':
          description: File is too large (maximum 500MB)
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: 'File is too large. Maximum size is 500MB'
        '500':
          description: Server error during processing
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
...
  /api/entityTypes:
    get:
      tags:
        - Queries
      summary: List entity types
      description: Returns all entity types available in specified models
      operationId: listEntityTypes
      parameters:
        - name: models
          in: query
          description: Comma-separated list of model names (optional - returns types from all models if omitted)
          schema:
            type: string
            example: HelloWallIFCjsonC-2x3
      responses:
        '200':
          description: List of entity types
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                example:
                  - IfcPropertySet
                  - IfcWall
                  - IfcRelDefinesByProperties
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /api/entityGuids:
    get:
      tags:
        - Queries
      summary: Query for entity GUIDs
      description: |
        Query entity GUIDs based on optional filters.
        Results are organized by model name.
        - Filter by model name(s) (optional)
        - Filter by entity type(s) (optional)
      operationId: queryEntityGuids
      parameters:
        - name: models
          in: query
          description: Comma-separated list of model names to filter by (optional - includes all models if omitted)
          schema:
            type: string
            example: HelloWallIFCjsonC-2x3
        - name: entityTypes
          in: query
          description: Comma-separated list of entity types to filter by
          schema:
            type: string
            example: IfcWall,IfcPropertySet
      responses:
        '200':
          description: Dictionary mapping model names to arrays of entity GUIDs
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: array
                  items:
                    type: string
                example:
                  HelloWallIFCjsonC-2x3:
                    - entity-guid-1
                    - entity-guid-2
                    - entity-guid-3
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /api/componentGuids:
    get:
      tags:
        - Queries
      summary: Query for component GUIDs
      description: |
        Query component GUIDs based on optional filters.
        Results are organized by model name.
        - Filter by model name(s) (optional)
        - Filter by entity GUID(s) (optional)
        - Filter by entity type(s) (optional)
      operationId: queryComponentGuids
      parameters:
        - name: models
          in: query
          description: Comma-separated list of model names to filter by (optional - includes all models if omitted)
          schema:
            type: string
            example: HelloWallIFCjsonC-2x3
        - name: entityGuids
          in: query
          description: Comma-separated list of entity GUIDs to filter components by
          schema:
            type: string
            example: 'entity-guid-1,entity-guid-2'
        - name: entityTypes
          in: query
          description: Comma-separated list of entity types to filter by
          schema:
            type: string
            example: IfcWall
      responses:
        '200':
          description: Dictionary mapping model names to arrays of component GUIDs
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: array
                  items:
                    type: string
                example:
                  HelloWallIFCjsonC-2x3:
                    - component-guid-1
                    - component-guid-2
                    - component-guid-3
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /api/components:
    get:
      tags:
        - Queries
      summary: Retrieve full component data with flexible filtering
      description: |
        Retrieve complete component data (including all properties and metadata).
        Results are organized by model name. Supports flexible filtering:
        - Filter by specific component GUIDs (optional)
        - Filter by model name(s) (optional)
        - Filter by entity type(s) (optional)
        - Filter by entity GUID(s) (optional)
      operationId: getComponents
      parameters:
        - name: componentGuids
          in: query
          description: Comma-separated list of specific component GUIDs to retrieve (optional)
          schema:
            type: string
            example: 'component-guid-1,component-guid-2'
        - name: models
          in: query
          description: Comma-separated list of model names to filter by (optional)
          schema:
            type: string
            example: HelloWall-01,HelloWall-02
        - name: entityTypes
          in: query
          description: Comma-separated list of entity types to filter by (optional)
          schema:
            type: string
            example: IfcWallStandardCase
        - name: entityGuids
          in: query
          description: Comma-separated list of entity GUIDs to filter by (optional)
          schema:
            type: string
            example: 'entity-guid-1,entity-guid-2'
      responses:
        '200':
          description: Dictionary mapping model names to arrays of component objects
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: array
                  items:
                    type: object
                    additionalProperties: true
                example:
                  HelloWall-01:
                    - componentGuid: component-guid-1
                      entityGuid: entity-guid-1
                      componentType: IfcWallStandardCaseComponent
                      entityType: IfcWallStandardCase
                      model: HelloWall-01
                      properties: {}
                    - componentGuid: component-guid-2
                      entityGuid: entity-guid-1
                      componentType: IfcWallStandardCaseComponent
                      entityType: IfcWallStandardCase
                      model: HelloWall-01
                      properties: {}
                  HelloWall-02:
                    - componentGuid: component-guid-3
                      entityGuid: entity-guid-2
                      componentType: IfcWallStandardCaseComponent
                      entityType: IfcWallStandardCase
                      model: HelloWall-02
                      properties: {}
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /api/refresh:
    post:
      tags:
        - Data Management
      summary: Manually refresh the in-memory component tree
      description: |
        Triggers a full refresh of the in-memory component tree from the file-based data store.
        This is useful after manually adding files to the data store directory.
      operationId: refreshMemory
      responses:
        '200':
          description: Memory tree successfully refreshed
          content:
            application/json:
              schema:
                type: object
                properties:
                  models_loaded:
                    type: integer
                    example: 1
                  message:
                    type: string
                    example: 'Memory tree refreshed with 1 model(s)'
        '500':
          description: Server error during refresh
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
...